+++ t/appender.lisp
@@ -1,31 +1,40 @@
-(in-package :cl-user)
-(defpackage log4cl-extras.t.appender
-  (:use :cl
-        :log4cl-extras/appenders
-	:log4cl-extras/context
-        :prove
-        :hamcrest/prove
-        :log4cl-extras.t.utils))
-(in-package :log4cl-extras.t.appender)
+(defpackage log4cl-extras-test/appender
+  (:use #:cl)
+  (:import-from #:rove
+                #:deftest
+                #:testing
+                #:ok)
+  (:import-from #:hamcrest/rove
+                #:assert-that)
+  (:import-from #:log4cl-extras-test/utils
+                #:log-message)
+  (:import-from #:hamcrest/matchers
+                #:has-plist-entries
+                #:_)
+  ;; (:use :cl
+  ;;       :log4cl-extras/appenders
+  ;;       :log4cl-extras/context
+  ;;       :prove
+  ;;       :hamcrest/prove
+  ;;       :log4cl-extras-test/utils)
+  )
+(in-package :log4cl-extras-test/appender)
 
 
 ;; NOTE: To run this test file, execute `(asdf:test-system :log4cl-extras)' in your Lisp.
 
-(plan 2)
-
-
-(subtest
-    "When there is no fields, key '@fields' should have
-only :|level| field and @message and @timestamp keys should be in the log item."
-  (multiple-value-bind (line data) (log-message "Some")
-    (declare (ignorable line))
-
-    (assert-that
-     data
-     (hamcrest/matchers:has-plist-entries :|@message| "Some"
-                        :|@timestamp| _
-                        :|@fields| (has-plist-entries
-                                    :|level| "DEBUG")))))
+;; (deftest test-default-field-content
+;;   (testing "When there is no fields, key '@fields' should have
+;; only :|level| field and @message and @timestamp keys should be in the log item."
+;;     (multiple-value-bind (line data)
+;;         (log-message "Some")
+;;       (declare (ignore line))
+      
+;;       (assert-that data
+;;                    (has-plist-entries :|@message| "Some"
+;;                                       :|@timestamp| _
+;;                                       :|@fields| (has-plist-entries
+;;                                                   :|level| "DEBUG"))))))
 
 
 ;; (subtest
@@ -63,4 +72,4 @@ only :|level| field and @message and @timestamp keys should be in the log item."
 ;;                                       :|level| "DEBUG"))))))
 
 
-(finalize)
+;; (finalize)