+++ t/core.lisp
@@ -1,21 +1,18 @@
-(in-package :cl-user)
-(defpackage log4cl-extras.t.core
-  (:use :cl
-        :log4cl-extras/error
-        :prove))
-(in-package :log4cl-extras.t.core)
+(defpackage log4cl-extras-test/core
+  (:use :cl))
+(in-package :log4cl-extras-test/core)
 
 
-(plan 1)
+;; (plan 1)
 
-(subtest
-    "Log unhandled."
-  (labels ((foo ()
-             (error "Blah minor"))
-           (bar ()
-             (foo)))
+;; (subtest
+;;     "Log unhandled."
+;;   (labels ((foo ()
+;;              (error "Blah minor"))
+;;            (bar ()
+;;              (foo)))
 
-    (multiple-value-bind (line data)
-        (with-log-unhandled ()
-            (bar)))))
-(finalize)
+;;     (multiple-value-bind (line data)
+;;         (with-log-unhandled ()
+;;             (bar)))))
+;; (finalize)