+++ t/utils.lisp
@@ -1,9 +1,9 @@
-(in-package :cl-user)
-(defpackage log4cl-extras.t.utils
-  (:use :cl
-        :log4cl-extras/appenders)
-  (:export :log-message))
-(in-package :log4cl-extras.t.utils)
+(defpackage log4cl-extras-test/utils
+  (:use #:cl
+        #:log4cl-extras/appenders)
+  (:import-from #:jonathan)
+  (:export #:log-message))
+(in-package :log4cl-extras-test/utils)
 
 
 (defun log-message (message &key (level log4cl:+log-level-debug+))